+++ pennylane_lightning/core/_serialize.py
@@ -25,6 +25,8 @@ from pennylane import (
     Identity,
     StatePrep,
     Rot,
+    Hamiltonian,
+    SparseHamiltonian,
 )
 from pennylane.operation import Tensor
 from pennylane.tape import QuantumTape
@@ -85,31 +87,25 @@ class QuantumScriptSerializer:
         self.tensor_prod_obs_c128 = lightning_ops.observables.TensorProdObsC128
         self.hamiltonian_c64 = lightning_ops.observables.HamiltonianC64
         self.hamiltonian_c128 = lightning_ops.observables.HamiltonianC128
-
-        if self.device_name == "lightning.gpu":
-            self.sparsehamiltonian_c64 = lightning_ops.observables.SparseHamiltonianC64
-            self.sparsehamiltonian_c128 = lightning_ops.observables.SparseHamiltonianC128
+        if device_name in ["lightning.gpu", "lightning.kokkos"]:
+            self.sparse_hamiltonian_c64 = lightning_ops.observables.SparseHamiltonianC64
+            self.sparse_hamiltonian_c128 = lightning_ops.observables.SparseHamiltonianC128
 
         self.use_mpi = False
 
         if use_mpi:
             self.use_mpi = use_mpi
-            self.statevectormpi_c128 = lightning_ops.StateVectorMPIC128
-            self.named_obsmpi_c64 = lightning_ops.observablesMPI.NamedObsMPIC64
-            self.named_obsmpi_c128 = lightning_ops.observablesMPI.NamedObsMPIC128
-            self.hermitian_obsmpi_c64 = lightning_ops.observablesMPI.HermitianObsMPIC64
-            self.hermitian_obsmpi_c128 = lightning_ops.observablesMPI.HermitianObsMPIC128
-            self.tensor_prod_obsmpi_c64 = lightning_ops.observablesMPI.TensorProdObsMPIC64
-            self.tensor_prod_obsmpi_c128 = lightning_ops.observablesMPI.TensorProdObsMPIC128
-            self.hamiltonianmpi_c64 = lightning_ops.observablesMPI.HamiltonianMPIC64
-            self.hamiltonianmpi_c128 = lightning_ops.observablesMPI.HamiltonianMPIC128
+            self.statevector_mpi_c128 = lightning_ops.StateVectorMPIC128
+            self.named_obs_mpi_c64 = lightning_ops.observablesMPI.NamedObsMPIC64
+            self.named_obs_mpi_c128 = lightning_ops.observablesMPI.NamedObsMPIC128
+            self.hermitian_obs_mpi_c64 = lightning_ops.observablesMPI.HermitianObsMPIC64
+            self.hermitian_obs_mpi_c128 = lightning_ops.observablesMPI.HermitianObsMPIC128
+            self.tensor_prod_obs_mpi_c64 = lightning_ops.observablesMPI.TensorProdObsMPIC64
+            self.tensor_prod_obs_mpi_c128 = lightning_ops.observablesMPI.TensorProdObsMPIC128
+            self.hamiltonian_mpi_c64 = lightning_ops.observablesMPI.HamiltonianMPIC64
+            self.hamiltonian_mpi_c128 = lightning_ops.observablesMPI.HamiltonianMPIC128
 
-            if self.device_name == "lightning.gpu":
-                self.sparsehamiltonianmpi_c64 = lightning_ops.observablesMPI.SparseHamiltonianMPIC64
-                self.sparsehamiltonianmpi_c128 = (
-                    lightning_ops.observablesMPI.SparseHamiltonianMPIC128
-                )
-            self.mpi_manager = lightning_ops.MPIManager
+            self._mpi_manager = lightning_ops.MPIManager
 
     @property
     def ctype(self):
@@ -124,47 +120,43 @@ class QuantumScriptSerializer:
     @property
     def sv_type(self):
         if self.use_mpi:
-            return self.statevectormpi_c128
+            return self.statevector_mpi_c128
         return self.statevector_c128
 
     @property
     def named_obs(self):
         """Named observable matching ``use_csingle`` precision."""
         if self.use_mpi:
-            return self.named_obsmpi_c64 if self.use_csingle else self.named_obsmpi_c128
+            return self.named_obs_mpi_c64 if self.use_csingle else self.named_obs_mpi_c128
         return self.named_obs_c64 if self.use_csingle else self.named_obs_c128
 
     @property
     def hermitian_obs(self):
         """Hermitian observable matching ``use_csingle`` precision."""
         if self.use_mpi:
-            return self.hermitian_obsmpi_c64 if self.use_csingle else self.hermitian_obsmpi_c128
+            return self.hermitian_obs_mpi_c64 if self.use_csingle else self.hermitian_obs_mpi_c128
         return self.hermitian_obs_c64 if self.use_csingle else self.hermitian_obs_c128
 
     @property
     def tensor_obs(self):
         """Tensor product observable matching ``use_csingle`` precision."""
         if self.use_mpi:
-            return self.tensor_prod_obsmpi_c64 if self.use_csingle else self.tensor_prod_obsmpi_c128
+            return (
+                self.tensor_prod_obs_mpi_c64 if self.use_csingle else self.tensor_prod_obs_mpi_c128
+            )
         return self.tensor_prod_obs_c64 if self.use_csingle else self.tensor_prod_obs_c128
 
     @property
     def hamiltonian_obs(self):
         """Hamiltonian observable matching ``use_csingle`` precision."""
         if self.use_mpi:
-            return self.hamiltonianmpi_c64 if self.use_csingle else self.hamiltonianmpi_c128
+            return self.hamiltonian_mpi_c64 if self.use_csingle else self.hamiltonian_mpi_c128
         return self.hamiltonian_c64 if self.use_csingle else self.hamiltonian_c128
 
     @property
-    def sparsehamiltonian_obs(self):
-        """Sparse Hamiltonian observable matching ``use_csingle`` precision."""
-        if self.use_mpi:
-            return (
-                self.sparsehamiltonianmpi_c64
-                if self.use_csingle
-                else self.sparsehamiltonianmpi_c128
-            )
-        return self.sparsehamiltonian_c64 if self.use_csingle else self.sparsehamiltonian_c128
+    def sparse_hamiltonian_obs(self):
+        """SparseHamiltonian observable matching ``use_csingle`` precision."""
+        return self.sparse_hamiltonian_c64 if self.use_csingle else self.sparse_hamiltonian_c128
 
     def _named_obs(self, observable, wires_map: dict):
         """Serializes a Named observable"""
@@ -190,25 +182,37 @@ class QuantumScriptSerializer:
         terms = [self._ob(t, wires_map) for t in observable.ops]
         return self.hamiltonian_obs(coeffs, terms)
 
-    def _sparsehamiltonian(self, observable, wires_map: dict):
+    def _sparse_hamiltonian(self, observable, wires_map: dict):
+        """Serialize an observable (Sparse Hamiltonian)
+
+        Args:
+            observable (Observable): the input observable (Sparse Hamiltonian)
+            wire_map (dict): a dictionary mapping input wires to the device's backend wires
+
+        Returns:
+            sparse_hamiltonian_obs (SparseHamiltonianC64 or SparseHamiltonianC128): A Sparse Hamiltonian observable object compatible with the C++ backend
+        """
+
+        if self.use_mpi:
+            Hmat = Hamiltonian([1.0], [Identity(0)]).sparse_matrix()
+            H_sparse = SparseHamiltonian(Hmat, wires=range(1))
+            spm = H_sparse.sparse_matrix()
+            # Only root 0 needs the overall sparsematrix data
+            if self._mpi_manager().getRank() == 0:
+                spm = observable.sparse_matrix()
+            self._mpi_manager().Barrier()
+        else:
+            spm = observable.sparse_matrix()
+        spm = observable.sparse_matrix()
+        data = np.array(spm.data).astype(self.ctype)
+        indices = np.array(spm.indices).astype(np.int64)
+        offsets = np.array(spm.indptr).astype(np.int64)
+
         wires = []
         wires_list = observable.wires.tolist()
         wires.extend([wires_map[w] for w in wires_list])
-        if self.use_mpi:
-            Hmat = qml.Hamiltonian([1.0], [qml.Identity(0)]).sparse_matrix()
-            H_sparse = qml.SparseHamiltonian(Hmat, wires=range(1))
-            spm = H_sparse.sparse_matrix()
-            # Only root 0 needs the overall sparsematrix data
-            if self.mpi_manager().getRank() == 0:
-                spm = observable.sparse_matrix()
-            self.mpi_manager().Barrier()
-        else:
-            spm = observable.sparse_matrix()
-        data = np.array(spm.data).astype(self.ctype)
-        indices = np.array(spm.indices).astype(self.rtype)
-        offsets = np.array(spm.indptr).astype(self.rtype)
 
-        return self.sparsehamiltonian_obs(data, indices, offsets, wires)
+        return self.sparse_hamiltonian_obs(data, indices, offsets, wires)
 
     def _pauli_word(self, observable, wires_map: dict):
         """Serialize a :class:`pennylane.pauli.PauliWord` into a Named or Tensor observable."""
@@ -237,8 +241,11 @@ class QuantumScriptSerializer:
             return self._tensor_ob(observable, wires_map)
         if observable.name == "Hamiltonian":
             return self._hamiltonian(observable, wires_map)
-        if observable.name == "SparseHamiltonian" and self.device_name != "lightning.qubit":
-            return self._sparsehamiltonian(observable, wires_map)
+        if (
+            self.device_name in ["lightning.gpu", "lightning.kokkos"]
+            and observable.name == "SparseHamiltonian"
+        ):
+            return self._sparse_hamiltonian(observable, wires_map)
         if isinstance(observable, (PauliX, PauliY, PauliZ, Identity, Hadamard)):
             return self._named_obs(observable, wires_map)
         if observable._pauli_rep is not None: