+++ pennylane_lightning/core/src/observables/Observables.hpp
@@ -414,6 +414,7 @@ class HamiltonianBase : public Observable<StateVectorT> {
     }
 };
 
+// NOLINTBEGIN
 /**
  * @brief Sparse representation of SparseHamiltonian<T>
  *
@@ -424,8 +425,8 @@ class SparseHamiltonianBase : public Observable<StateVectorT> {
   public:
     using PrecisionT = typename StateVectorT::PrecisionT;
     using ComplexT = typename StateVectorT::ComplexT;
-    // cuSparse required index type
 #ifdef _ENABLE_PLGPU
+    // cuSparse required index type
     using IdxT =
         typename std::conditional<std::is_same<PrecisionT, float>::value,
                                   int32_t, int64_t>::type;
@@ -480,7 +481,7 @@ class SparseHamiltonianBase : public Observable<StateVectorT> {
      *
      * @param arg1 Argument to construct data
      * @param arg2 Argument to construct indices
-     * @param arg3 Argument to construct ofsets
+     * @param arg3 Argument to construct offsets
      * @param arg4 Argument to construct wires
      */
     static auto create(std::initializer_list<ComplexT> arg1,
@@ -504,18 +505,20 @@ class SparseHamiltonianBase : public Observable<StateVectorT> {
         using Pennylane::Util::operator<<;
         std::ostringstream ss;
         ss << "SparseHamiltonian: {\n'data' : \n";
-        for (const auto &d : data_)
-            // Note that for LGPU backend, ComplexT is std::complex as of 0.33
+        for (const auto &d : data_) { // Note that for LGPU backend, ComplexT is
+                                      // std::complex as of 0.33
             // release Need to revisit it once we set ComplexT as cuComplex
             // later
-            ss << "{" << d.real() << ", " << d.imag() << "},"
-               << "\n";
+            ss << "{" << d.real() << ", " << d.imag() << "}, ";
+        }
         ss << ",\n'indices' : \n";
-        for (const auto &i : indices_)
-            ss << i << std::endl;
+        for (const auto &i : indices_) {
+            ss << i << ", ";
+        }
         ss << ",\n'offsets' : \n";
-        for (const auto &o : offsets_)
-            ss << o;
+        for (const auto &o : offsets_) {
+            ss << o << ", ";
+        }
         ss << "\n}";
         return ss.str();
     }
@@ -526,5 +529,6 @@ class SparseHamiltonianBase : public Observable<StateVectorT> {
         return wires_;
     };
 };
+// NOLINTEND
 
 } // namespace Pennylane::Observables